# lib/api-spec/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api",
    "version": "1.0.0",
    "description": "LIV LLC accreditation and certificate API"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/public/stats": {
      "get": {
        "operationId": "getPublicStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/verify/{certNumber}": {
      "get": {
        "operationId": "verifyCertificate",
        "parameters": [
          {
            "name": "certNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "t",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "last_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/v1/certificates": {
      "post": {
        "operationId": "apiIssueCertificates",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssuanceInput"
              }
            }
          }
        }
      }
    },
    "/v1/certificates/{certNumber}": {
      "get": {
        "operationId": "apiGetCertificate",
        "parameters": [
          {
            "name": "certNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/certificates/{certNumber}/revoke": {
      "post": {
        "operationId": "apiRevokeCertificate",
        "parameters": [
          {
            "name": "certNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevocationInput"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "JsonObject": {
        "type": "object",
        "additionalProperties": true
      },
...
      "RevocationInput": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 3
          }
        }
      },
...
      "TraineeInput": {
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "email",
          "course_name",
          "completion_date"
        ],
        "properties": {
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "course_name": {
            "type": "string"
          },
          "completion_date": {
            "type": "string"
          },
          "grade": {
            "type": "string"
          }
        }
      },
      "IssuanceInput": {
        "type": "object",
        "required": [
          "template_id",
          "rows"
        ],
        "properties": {
          "template_id": {
            "type": "integer"
          },
          "file_name": {
            "type": "string"
          },
          "skip_invalid": {
            "type": "boolean"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraineeInput"
            }
          }
        }
      },
...
    }
  }
}